+++ src/runtime/runtime.cs
@@ -1,8 +1,7 @@
-using System.Reflection.Emit;
 using System;
+using System.Diagnostics;
 using System.Diagnostics.Contracts;
 using System.Runtime.InteropServices;
-using System.Security;
 using System.Text;
 using System.Threading;
 using System.Collections.Generic;
@@ -86,9 +85,9 @@ namespace Python.Runtime
             {
                 using (var versionTuple = new PyTuple(PySys_GetObject("version_info")))
                 {
-                    var major = versionTuple[0].As<int>();
-                    var minor = versionTuple[1].As<int>();
-                    var micro = versionTuple[2].As<int>();
+                    var major = Converter.ToInt32(versionTuple[0].Reference);
+                    var minor = Converter.ToInt32(versionTuple[1].Reference);
+                    var micro = Converter.ToInt32(versionTuple[2].Reference);
                     return new Version(major, minor, micro);
                 }
             }
@@ -199,15 +198,15 @@ namespace Python.Runtime
                 SetPyMember(ref PyFalse, PyObject_GetAttrString(builtins, "False"),
                     () => PyFalse = IntPtr.Zero);
 
-                SetPyMember(ref PyBoolType, PyObject_Type(PyTrue),
+                SetPyMemberTypeOf(ref PyBoolType, PyTrue,
                     () => PyBoolType = IntPtr.Zero);
-                SetPyMember(ref PyNoneType, PyObject_Type(PyNone),
+                SetPyMemberTypeOf(ref PyNoneType, PyNone,
                     () => PyNoneType = IntPtr.Zero);
-                SetPyMember(ref PyTypeType, PyObject_Type(PyNoneType),
+                SetPyMemberTypeOf(ref PyTypeType, PyNoneType,
                     () => PyTypeType = IntPtr.Zero);
 
                 op = PyObject_GetAttrString(builtins, "len");
-                SetPyMember(ref PyMethodType, PyObject_Type(op),
+                SetPyMemberTypeOf(ref PyMethodType, op,
                     () => PyMethodType = IntPtr.Zero);
                 XDecref(op);
 
@@ -216,7 +215,7 @@ namespace Python.Runtime
                 //
                 // object.__init__ seems safe, though.
                 op = PyObject_GetAttr(PyBaseObjectType, PyIdentifier.__init__);
-                SetPyMember(ref PyWrapperDescriptorType, PyObject_Type(op),
+                SetPyMemberTypeOf(ref PyWrapperDescriptorType, op,
                     () => PyWrapperDescriptorType = IntPtr.Zero);
                 XDecref(op);
 
@@ -227,47 +226,47 @@ namespace Python.Runtime
             }
 
             op = PyString_FromString("string");
-            SetPyMember(ref PyStringType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyStringType, op,
                 () => PyStringType = IntPtr.Zero);
             XDecref(op);
 
             op = PyUnicode_FromString("unicode");
-            SetPyMember(ref PyUnicodeType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyUnicodeType, op,
                 () => PyUnicodeType = IntPtr.Zero);
             XDecref(op);
 
             op = EmptyPyBytes();
-            SetPyMember(ref PyBytesType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyBytesType, op,
                 () => PyBytesType = IntPtr.Zero);
             XDecref(op);
 
             op = PyTuple_New(0);
-            SetPyMember(ref PyTupleType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyTupleType, op,
                 () => PyTupleType = IntPtr.Zero);
             XDecref(op);
 
             op = PyList_New(0);
-            SetPyMember(ref PyListType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyListType, op,
                 () => PyListType = IntPtr.Zero);
             XDecref(op);
 
             op = PyDict_New();
-            SetPyMember(ref PyDictType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyDictType, op,
                 () => PyDictType = IntPtr.Zero);
             XDecref(op);
 
             op = PyInt_FromInt32(0);
-            SetPyMember(ref PyIntType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyIntType, op,
                 () => PyIntType = IntPtr.Zero);
             XDecref(op);
 
             op = PyLong_FromLong(0);
-            SetPyMember(ref PyLongType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyLongType, op,
                 () => PyLongType = IntPtr.Zero);
             XDecref(op);
 
             op = PyFloat_FromDouble(0);
-            SetPyMember(ref PyFloatType, PyObject_Type(op),
+            SetPyMemberTypeOf(ref PyFloatType, op,
                 () => PyFloatType = IntPtr.Zero);
             XDecref(op);
 
@@ -279,7 +278,8 @@ namespace Python.Runtime
             _PyObject_NextNotImplemented = Get_PyObject_NextNotImplemented();
             {
                 using var sys = PyImport_ImportModule("sys");
-                PyModuleType = PyObject_Type(sys.DangerousMoveToPointer());
+                SetPyMemberTypeOf(ref PyModuleType, sys.DangerousGetAddress(),
+                    () => PyModuleType = IntPtr.Zero);
             }
         }
 
@@ -456,6 +456,12 @@ namespace Python.Runtime
             _pyRefs.Add(value, onRelease);
         }
 
+        private static void SetPyMemberTypeOf(ref IntPtr obj, IntPtr value, Action onRelease)
+        {
+            var type = PyObject_Type(new BorrowedReference(value)).DangerousMoveToPointer();
+            SetPyMember(ref obj, type, onRelease);
+        }
+
         private static void ResetPyMembers()
         {
             _pyRefs.Release();
@@ -471,7 +477,7 @@ namespace Python.Runtime
                 var item = PyList_GetItem(items, i);
                 var name = PyTuple_GetItem(item, 0);
                 var module = PyTuple_GetItem(item, 1);
-                if (ManagedType.IsManagedType(module))
+                if (ManagedType.IsInstanceOfManagedType(module))
                 {
                     PyDict_DelItem(modules, name);
                 }
@@ -524,6 +530,7 @@ namespace Python.Runtime
                 if (obj.gcHandle.IsAllocated)
                 {
                     obj.gcHandle.Free();
+                    ManagedType.SetGCHandle(obj.ObjectReference, default);
                 }
                 obj.gcHandle = default;
             }
@@ -568,6 +575,8 @@ namespace Python.Runtime
         internal static IntPtr PyFalse;
         internal static IntPtr PyNone;
         internal static IntPtr Error;
+
+        internal static BorrowedReference CLRMetaType => new BorrowedReference(PyCLRMetaType);
 
         public static PyObject None
         {
@@ -759,16 +768,12 @@ namespace Python.Runtime
         [Pure]
         internal static unsafe long Refcount(IntPtr op)
         {
-#if PYTHON_WITH_PYDEBUG
-            var p = (void*)(op + TypeOffset.ob_refcnt);
-#else
-            var p = (void*)op;
-#endif
-            if ((void*)0 == p)
+            if (op == IntPtr.Zero)
             {
                 return 0;
             }
-            return Is32Bit ? (*(int*)p) : (*(long*)p);
+            var p = (nint*)(op + ABI.RefCountOffset);
+            return *p;
         }
 
         /// <summary>
@@ -975,14 +980,9 @@ namespace Python.Runtime
             {
                 return IntPtr.Zero;
             }
-#if PYTHON_WITH_PYDEBUG
-            var n = 3;
-#else
-            var n = 1;
-#endif
-            return Is32Bit
-                ? new IntPtr((void*)(*((uint*)p + n)))
-                : new IntPtr((void*)(*((ulong*)p + n)));
+            Debug.Assert(TypeOffset.ob_type > 0);
+            IntPtr* typePtr = (IntPtr*)(op + TypeOffset.ob_type);
+            return *typePtr;
         }
         internal static unsafe BorrowedReference PyObject_TYPE(BorrowedReference op)
             => new BorrowedReference(PyObject_TYPE(op.DangerousGetAddress()));
@@ -999,9 +999,12 @@ namespace Python.Runtime
             return tp;
         }
 
+        internal static NewReference PyObject_Type(BorrowedReference o)
+            => Delegates.PyObject_Type(o);
+
         internal static string PyObject_GetTypeName(IntPtr op)
         {
-            IntPtr pyType = Marshal.ReadIntPtr(op, ObjectOffset.ob_type);
+            IntPtr pyType = PyObject_TYPE(op);
             IntPtr ppName = Marshal.ReadIntPtr(pyType, TypeOffset.tp_name);
             return Marshal.PtrToStringAnsi(ppName);
         }
@@ -1011,7 +1014,7 @@ namespace Python.Runtime
         /// </summary>
         internal static bool PyObject_IsIterable(IntPtr pointer)
         {
-            var ob_type = Marshal.ReadIntPtr(pointer, ObjectOffset.ob_type);
+            var ob_type = PyObject_TYPE(pointer);
             IntPtr tp_iter = Marshal.ReadIntPtr(ob_type, TypeOffset.tp_iter);
             return tp_iter != IntPtr.Zero;
         }
@@ -1109,7 +1112,7 @@ namespace Python.Runtime
         internal static int PyObject_IsInstance(IntPtr ob, IntPtr type) => Delegates.PyObject_IsInstance(ob, type);
 
 
-        internal static int PyObject_IsSubclass(IntPtr ob, IntPtr type) => Delegates.PyObject_IsSubclass(ob, type);
+        internal static int PyObject_IsSubclass(BorrowedReference ob, BorrowedReference type) => Delegates.PyObject_IsSubclass(ob, type);
 
 
         internal static int PyCallable_Check(IntPtr pointer) => Delegates.PyCallable_Check(pointer);
@@ -1127,21 +1130,27 @@ namespace Python.Runtime
         internal static nint PyObject_Hash(IntPtr op) => Delegates.PyObject_Hash(op);
 
 
-        internal static IntPtr PyObject_Repr(IntPtr pointer) => Delegates.PyObject_Repr(pointer);
+        internal static IntPtr PyObject_Repr(IntPtr pointer)
+        {
+            Debug.Assert(PyErr_Occurred() == IntPtr.Zero);
+            return Delegates.PyObject_Repr(pointer);
+        }
 
 
-        internal static IntPtr PyObject_Str(IntPtr pointer) => Delegates.PyObject_Str(pointer);
-
-
-        internal static IntPtr PyObject_Unicode(IntPtr pointer) => Delegates.PyObject_Unicode(pointer);
+        internal static IntPtr PyObject_Str(IntPtr pointer)
+        {
+            Debug.Assert(PyErr_Occurred() == IntPtr.Zero);
+            return Delegates.PyObject_Str(pointer);
+        }
 
 
         internal static IntPtr PyObject_Dir(IntPtr pointer) => Delegates.PyObject_Dir(pointer);
 
-#if PYTHON_WITH_PYDEBUG
-        [DllImport(_PythonDll, CallingConvention = CallingConvention.Cdecl)]
-        internal static extern void _Py_NewReference(IntPtr ob);
-#endif
+        internal static void _Py_NewReference(BorrowedReference ob)
+        {
+            if (Delegates._Py_NewReference != null)
+                Delegates._Py_NewReference(ob);
+        }
 
         //====================================================================
         // Python buffer API
@@ -1504,6 +1513,13 @@ namespace Python.Runtime
             return Delegates.PyBytes_FromString((IntPtr)bytes);
         }
 
+        internal static IntPtr PyBytes_AsString(IntPtr ob) => PyBytes_AsString(new BorrowedReference(ob));
+        internal static IntPtr PyBytes_AsString(BorrowedReference ob)
+        {
+            Debug.Assert(ob != null);
+            return Delegates.PyBytes_AsString(ob);
+        }
+
         internal static long PyBytes_Size(IntPtr op)
         {
             return (long)_PyBytes_Size(op);
@@ -1511,11 +1527,6 @@ namespace Python.Runtime
 
 
         private static IntPtr _PyBytes_Size(IntPtr op) => Delegates._PyBytes_Size(op);
-
-        internal static IntPtr PyBytes_AS_STRING(IntPtr ob)
-        {
-            return ob + BytesOffset.ob_sval;
-        }
 
 
         internal static IntPtr PyUnicode_FromStringAndSize(IntPtr value, long size)
@@ -1611,7 +1622,7 @@ namespace Python.Runtime
             {
                 using var p = PyUnicode_AsUTF16String(new BorrowedReference(op));
                 int length = (int)PyUnicode_GetSize(op);
-                char* codePoints = (char*)PyBytes_AS_STRING(p.DangerousGetAddress());
+                char* codePoints = (char*)PyBytes_AsString(p.DangerousGetAddress());
                 return new string(codePoints,
                                   startIndex: 1, // skip BOM
                                   length: length);
@@ -1872,7 +1883,7 @@ namespace Python.Runtime
 
         internal static bool PyIter_Check(IntPtr pointer)
         {
-            var ob_type = Marshal.ReadIntPtr(pointer, ObjectOffset.ob_type);
+            var ob_type = PyObject_TYPE(pointer);
             IntPtr tp_iternext = Marshal.ReadIntPtr(ob_type, TypeOffset.tp_iternext);
             return tp_iternext != IntPtr.Zero && tp_iternext != _PyObject_NextNotImplemented;
         }
@@ -1903,11 +1914,6 @@ namespace Python.Runtime
         internal static string PyModule_GetFilename(IntPtr module)
             => Delegates.PyModule_GetFilename(module).ToString(Encoding.UTF8);
 
-#if PYTHON_WITH_PYDEBUG
-        [DllImport(_PythonDll, EntryPoint = "PyModule_Create2TraceRefs", CallingConvention = CallingConvention.Cdecl)]
-#else
-
-#endif
         internal static IntPtr PyModule_Create2(IntPtr module, int apiver) => Delegates.PyModule_Create2(module, apiver);
 
 
@@ -2006,6 +2012,8 @@ namespace Python.Runtime
 
 
         private static IntPtr PyType_GenericAlloc(IntPtr type, IntPtr n) => Delegates.PyType_GenericAlloc(type, n);
+        
+        internal static NewReference PyType_FromSpecWithBases(in NativeTypeSpec spec, BorrowedReference bases) => Delegates.PyType_FromSpecWithBases(in spec, bases);
 
         /// <summary>
         /// Finalize a type object. This should be called on all type objects to finish their initialization. This function is responsible for adding inherited slots from a type’s base class. Return 0 on success, or return -1 and sets an exception on error.
@@ -2242,7 +2250,7 @@ namespace Python.Runtime
             return PyImport_Import(PyIdentifier.builtins);
         }
 
-        private static class Delegates
+        internal static class Delegates
         {
             static readonly ILibraryLoader libraryLoader = LibraryLoader.Instance;
 
@@ -2312,7 +2320,7 @@ namespace Python.Runtime
                 PyObject_CallObject = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr, IntPtr>)GetFunctionByName(nameof(PyObject_CallObject), GetUnmanagedDll(_PythonDll));
                 PyObject_RichCompareBool = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr, int, int>)GetFunctionByName(nameof(PyObject_RichCompareBool), GetUnmanagedDll(_PythonDll));
                 PyObject_IsInstance = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr, int>)GetFunctionByName(nameof(PyObject_IsInstance), GetUnmanagedDll(_PythonDll));
-                PyObject_IsSubclass = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr, int>)GetFunctionByName(nameof(PyObject_IsSubclass), GetUnmanagedDll(_PythonDll));
+                PyObject_IsSubclass = (delegate* unmanaged[Cdecl]<BorrowedReference, BorrowedReference, int>)GetFunctionByName(nameof(PyObject_IsSubclass), GetUnmanagedDll(_PythonDll));
                 PyCallable_Check = (delegate* unmanaged[Cdecl]<IntPtr, int>)GetFunctionByName(nameof(PyCallable_Check), GetUnmanagedDll(_PythonDll));
                 PyObject_IsTrue = (delegate* unmanaged[Cdecl]<BorrowedReference, int>)GetFunctionByName(nameof(PyObject_IsTrue), GetUnmanagedDll(_PythonDll));
                 PyObject_Not = (delegate* unmanaged[Cdecl]<IntPtr, int>)GetFunctionByName(nameof(PyObject_Not), GetUnmanagedDll(_PythonDll));
@@ -2320,7 +2328,7 @@ namespace Python.Runtime
                 PyObject_Hash = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName(nameof(PyObject_Hash), GetUnmanagedDll(_PythonDll));
                 PyObject_Repr = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName(nameof(PyObject_Repr), GetUnmanagedDll(_PythonDll));
                 PyObject_Str = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName(nameof(PyObject_Str), GetUnmanagedDll(_PythonDll));
-                PyObject_Unicode = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName("PyObject_Str", GetUnmanagedDll(_PythonDll));
+                PyObject_Type = (delegate* unmanaged[Cdecl]<BorrowedReference, NewReference>)GetFunctionByName(nameof(PyObject_Type), GetUnmanagedDll(_PythonDll));
                 PyObject_Dir = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName(nameof(PyObject_Dir), GetUnmanagedDll(_PythonDll));
                 PyObject_GetBuffer = (delegate* unmanaged[Cdecl]<IntPtr, ref Py_buffer, int, int>)GetFunctionByName(nameof(PyObject_GetBuffer), GetUnmanagedDll(_PythonDll));
                 PyBuffer_Release = (delegate* unmanaged[Cdecl]<ref Py_buffer, void>)GetFunctionByName(nameof(PyBuffer_Release), GetUnmanagedDll(_PythonDll));
@@ -2401,6 +2409,7 @@ namespace Python.Runtime
                 _PySequence_Count = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr, IntPtr>)GetFunctionByName("PySequence_Count", GetUnmanagedDll(_PythonDll));
                 PySequence_Tuple = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName(nameof(PySequence_Tuple), GetUnmanagedDll(_PythonDll));
                 PySequence_List = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName(nameof(PySequence_List), GetUnmanagedDll(_PythonDll));
+                PyBytes_AsString = (delegate* unmanaged[Cdecl]<BorrowedReference, IntPtr>)GetFunctionByName(nameof(PyBytes_AsString), GetUnmanagedDll(_PythonDll));
                 PyBytes_FromString = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName(nameof(PyBytes_FromString), GetUnmanagedDll(_PythonDll));
                 _PyBytes_Size = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName("PyBytes_Size", GetUnmanagedDll(_PythonDll));
                 PyUnicode_FromStringAndSize = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr, IntPtr>)GetFunctionByName(nameof(PyUnicode_FromStringAndSize), GetUnmanagedDll(_PythonDll));
@@ -2455,7 +2464,14 @@ namespace Python.Runtime
                 PyModule_GetName = (delegate* unmanaged[Cdecl]<IntPtr, StrPtr>)GetFunctionByName(nameof(PyModule_GetName), GetUnmanagedDll(_PythonDll));
                 PyModule_GetDict = (delegate* unmanaged[Cdecl]<BorrowedReference, BorrowedReference>)GetFunctionByName(nameof(PyModule_GetDict), GetUnmanagedDll(_PythonDll));
                 PyModule_GetFilename = (delegate* unmanaged[Cdecl]<IntPtr, StrPtr>)GetFunctionByName(nameof(PyModule_GetFilename), GetUnmanagedDll(_PythonDll));
-                PyModule_Create2 = (delegate* unmanaged[Cdecl]<IntPtr, int, IntPtr>)GetFunctionByName(nameof(PyModule_Create2), GetUnmanagedDll(_PythonDll));
+                try
+                {
+                    PyModule_Create2 = (delegate* unmanaged[Cdecl]<IntPtr, int, IntPtr>)GetFunctionByName(nameof(PyModule_Create2), GetUnmanagedDll(_PythonDll));
+                }
+                catch (MissingMethodException)
+                {
+                    PyModule_Create2 = (delegate* unmanaged[Cdecl]<IntPtr, int, IntPtr>)GetFunctionByName("PyModule_Create2TraceRefs", GetUnmanagedDll(_PythonDll));
+                }
                 PyImport_Import = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr>)GetFunctionByName(nameof(PyImport_Import), GetUnmanagedDll(_PythonDll));
                 PyImport_ImportModule = (delegate* unmanaged[Cdecl]<StrPtr, NewReference>)GetFunctionByName(nameof(PyImport_ImportModule), GetUnmanagedDll(_PythonDll));
                 PyImport_ReloadModule = (delegate* unmanaged[Cdecl]<BorrowedReference, NewReference>)GetFunctionByName(nameof(PyImport_ReloadModule), GetUnmanagedDll(_PythonDll));
@@ -2509,6 +2525,13 @@ namespace Python.Runtime
                 PyException_SetCause = (delegate* unmanaged[Cdecl]<IntPtr, IntPtr, void>)GetFunctionByName(nameof(PyException_SetCause), GetUnmanagedDll(_PythonDll));
                 PyThreadState_SetAsyncExcLLP64 = (delegate* unmanaged[Cdecl]<uint, IntPtr, int>)GetFunctionByName("PyThreadState_SetAsyncExc", GetUnmanagedDll(_PythonDll));
                 PyThreadState_SetAsyncExcLP64 = (delegate* unmanaged[Cdecl]<ulong, IntPtr, int>)GetFunctionByName("PyThreadState_SetAsyncExc", GetUnmanagedDll(_PythonDll));
+                PyType_FromSpecWithBases = (delegate* unmanaged[Cdecl]<in NativeTypeSpec, BorrowedReference, NewReference>)GetFunctionByName(nameof(PyType_FromSpecWithBases), GetUnmanagedDll(PythonDLL));
+
+                try
+                {
+                    _Py_NewReference = (delegate* unmanaged[Cdecl]<BorrowedReference, void>)GetFunctionByName(nameof(_Py_NewReference), GetUnmanagedDll(_PythonDll));
+                }
+                catch (MissingMethodException) { }
             }
 
             static global::System.IntPtr GetUnmanagedDll(string libraryName)
@@ -2518,7 +2541,19 @@ namespace Python.Runtime
             }
 
             static global::System.IntPtr GetFunctionByName(string functionName, global::System.IntPtr libraryHandle)
-                => libraryLoader.GetFunction(libraryHandle, functionName);
+            {
+                try
+                {
+                    return libraryLoader.GetFunction(libraryHandle, functionName);
+                }
+                catch (MissingMethodException e) when (libraryHandle == IntPtr.Zero)
+                {
+                    throw new MissingMethodException(
+                        "Did you forget to set Runtime.PythonDLL?" +
+                        " See https://github.com/pythonnet/pythonnet#embedding-python-in-net",
+                        e);
+                }
+            }
 
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PyDictProxy_New { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, void> Py_IncRef { get; }
@@ -2584,7 +2619,7 @@ namespace Python.Runtime
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr, IntPtr> PyObject_CallObject { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr, int, int> PyObject_RichCompareBool { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr, int> PyObject_IsInstance { get; }
-            internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr, int> PyObject_IsSubclass { get; }
+            internal static delegate* unmanaged[Cdecl]<BorrowedReference, BorrowedReference, int> PyObject_IsSubclass { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, int> PyCallable_Check { get; }
             internal static delegate* unmanaged[Cdecl]<BorrowedReference, int> PyObject_IsTrue { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, int> PyObject_Not { get; }
@@ -2592,7 +2627,7 @@ namespace Python.Runtime
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PyObject_Hash { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PyObject_Repr { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PyObject_Str { get; }
-            internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PyObject_Unicode { get; }
+            internal static delegate* unmanaged[Cdecl]<BorrowedReference, NewReference> PyObject_Type { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PyObject_Dir { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, ref Py_buffer, int, int> PyObject_GetBuffer { get; }
             internal static delegate* unmanaged[Cdecl]<ref Py_buffer, void> PyBuffer_Release { get; }
@@ -2666,6 +2701,7 @@ namespace Python.Runtime
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr, IntPtr> _PySequence_Count { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PySequence_Tuple { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PySequence_List { get; }
+            internal static delegate* unmanaged[Cdecl]<BorrowedReference, IntPtr> PyBytes_AsString { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> PyBytes_FromString { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr> _PyBytes_Size { get; }
             internal static delegate* unmanaged[Cdecl]<IntPtr, IntPtr, IntPtr> PyUnicode_FromStringAndSize { get; }
@@ -2774,6 +2810,8 @@ namespace Python.Runtime
             internal static delegate* unmanaged[Cdecl]<uint, IntPtr, int> PyThreadState_SetAsyncExcLLP64 { get; }
             internal static delegate* unmanaged[Cdecl]<ulong, IntPtr, int> PyThreadState_SetAsyncExcLP64 { get; }
             internal static delegate* unmanaged[Cdecl]<BorrowedReference, IntPtr, NewReference> PyObject_GenericGetDict { get; }
+            internal static delegate* unmanaged[Cdecl]<in NativeTypeSpec, BorrowedReference, NewReference> PyType_FromSpecWithBases { get; }
+            internal static delegate* unmanaged[Cdecl]<BorrowedReference, void> _Py_NewReference { get; }
         }
     }